Face_recognition/requirement.txt
facenet-pytorch==2.6.0
gdown==5.2.0
gTTS==2.5.4
h5py==3.12.1
keras==3.6.0
Markdown==3.7
matplotlib==3.9.2
mtcnn==1.0.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.2.0
pip==24.2
playsound==1.2.2
pyttsx3==2.98
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
torch==2.2.2
torchaudio==2.5.1
torchvision==0.17.2
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
urllib3==2.2.3
customtkinter==5.2.2      
CustomTkinterMessagebox==0.0.4  
